+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+# [5.8.0] - 2024-03-25
+
+### Added
+
+* Added configurable endpoint overide for claim report
+
 # [5.7.0] - 2024-03-07
 
 ### Added
@@ -619,6 +625,7 @@ Added the ability to hit professional claim submission API. For more details, se
 * Authentication
 * Configuration
 
+[5.8.0]: https://github.com/WeInfuse/change_health/compare/v5.7.0...v5.8.0
 [5.7.0]: https://github.com/WeInfuse/change_health/compare/v5.6.0...v5.7.0
 [5.6.0]: https://github.com/WeInfuse/change_health/compare/v5.5.0...v5.6.0
 [5.5.0]: https://github.com/WeInfuse/change_health/compare/v5.4.2...v5.5.0